Add files via upload
Day 4: Not all the data is linear. when the data is nonlinear and one has to make the prediction with the least loss, then Polynomial Regression comes into play.
in this ipnyb file, i plotted a graph of nonlinear data first and then plotted the best-fit line using scikit learn.